Remove all empty columns on the right end of a data set.

This function will drop empty columns at the far right of a data set. This is useful when you import data and get empty columns on the right that shouldn't be there. This function only removes empty columns on the right - it doesn't remove empty columns anywhere else.

Usage

remove_right_empty_cols(data)

Arguments

data

a data frame

Value

the inputted data frame without empty columns on the right

Examples

data <- data.frame(
  col1 = sample(1:100, 100, replace = TRUE),
  col2 = NA,
  col3 = runif(100),
  col4 = sample(c("blue", "green", "yellow", "orange"), 100, replace = TRUE),
  col5 = NA,
  col6 = NA)

remove_right_empty_cols(data)
#>     col1 col2        col3   col4
#> 1     63   NA 0.332799758   blue
#> 2      4   NA 0.568352669   blue
#> 3     34   NA 0.252205725  green
#> 4     35   NA 0.464013567  green
#> 5     89   NA 0.917660507   blue
#> 6     86   NA 0.972844218   blue
#> 7     43   NA 0.819082447 yellow
#> 8      7   NA 0.902923798   blue
#> 9     32   NA 0.581366044 orange
#> 10    53   NA 0.773008481 yellow
#> 11    70   NA 0.995123026 orange
#> 12     9   NA 0.710971250 yellow
#> 13    22   NA 0.214942596 yellow
#> 14    10   NA 0.291757630 orange
#> 15    32   NA 0.721759729 orange
#> 16    34   NA 0.866615703 yellow
#> 17    17   NA 0.238453106   blue
#> 18    52   NA 0.004496308 orange
#> 19    56   NA 0.943516464   blue
#> 20    22   NA 0.438137200   blue
#> 21    95   NA 0.750603328   blue
#> 22     1   NA 0.667815764  green
#> 23    70   NA 0.407973201 yellow
#> 24    34   NA 0.351248815 yellow
#> 25    51   NA 0.738091561 orange
#> 26    46   NA 0.664285493  green
#> 27    65   NA 0.085224700 yellow
#> 28    25   NA 0.856132157 orange
#> 29    37   NA 0.076983322  green
#> 30     8   NA 0.852844803 yellow
#> 31    33   NA 0.106346961  green
#> 32    96   NA 0.484802824 yellow
#> 33    83   NA 0.247219110  green
#> 34    92   NA 0.686569211  green
#> 35    28   NA 0.163623198  green
#> 36    48   NA 0.952824800  green
#> 37    75   NA 0.321854551  green
#> 38    90   NA 0.361534117   blue
#> 39    35   NA 0.887723417 yellow
#> 40     2   NA 0.828014418 orange
#> 41    87   NA 0.100656458  green
#> 42     3   NA 0.906051578  green
#> 43    38   NA 0.772730364  green
#> 44    68   NA 0.383370670   blue
#> 45    65   NA 0.999652457   blue
#> 46     6   NA 0.349299049  green
#> 47    22   NA 0.947318266 yellow
#> 48    19   NA 0.216099976   blue
#> 49     2   NA 0.032092709 orange
#> 50    64   NA 0.145315843 orange
#> 51    40   NA 0.854383888  green
#> 52    65   NA 0.213149311  green
#> 53    28   NA 0.210310738 orange
#> 54    97   NA 0.039520690  green
#> 55    71   NA 0.944774803 orange
#> 56    12   NA 0.244927987 yellow
#> 57    54   NA 0.781122568  green
#> 58    41   NA 0.288237168 yellow
#> 59    42   NA 0.875357910 yellow
#> 60    66   NA 0.295750094 orange
#> 61    26   NA 0.983525408 orange
#> 62    60   NA 0.589837556 orange
#> 63    14   NA 0.759158380 yellow
#> 64     5   NA 0.836075306 orange
#> 65    76   NA 0.762819470 yellow
#> 66    17   NA 0.417269926  green
#> 67    13   NA 0.138074841   blue
#> 68    97   NA 0.080844962 orange
#> 69    33   NA 0.655982626  green
#> 70    77   NA 0.602003859 yellow
#> 71    82   NA 0.656995830   blue
#> 72    32   NA 0.329317164 orange
#> 73    69   NA 0.979474220 yellow
#> 74    55   NA 0.715186134 orange
#> 75    11   NA 0.872630305   blue
#> 76    33   NA 0.983283747   blue
#> 77    36   NA 0.218562993   blue
#> 78     7   NA 0.664530064   blue
#> 79    23   NA 0.389564038  green
#> 80    88   NA 0.046063639 yellow
#> 81    34   NA 0.616914558   blue
#> 82    72   NA 0.598474994 yellow
#> 83    73   NA 0.406853630 yellow
#> 84    51   NA 0.858328150  green
#> 85    59   NA 0.517681185 orange
#> 86    38   NA 0.979293410 yellow
#> 87    25   NA 0.017015688   blue
#> 88    53   NA 0.673447826 yellow
#> 89    35   NA 0.371269882   blue
#> 90    45   NA 0.918010637   blue
#> 91    82   NA 0.677978095 orange
#> 92    23   NA 0.665152461  green
#> 93    43   NA 0.756041088  green
#> 94     4   NA 0.542837155 yellow
#> 95    34   NA 0.239288098   blue
#> 96    41   NA 0.508893572  green
#> 97    82   NA 0.417264371  green
#> 98    94   NA 0.726948853 yellow
#> 99    19   NA 0.637685552 orange
#> 100   76   NA 0.396409955  green
 # the function drops `col5` and `col6` but keeps `col3`:
head(data)
#>   col1 col2      col3  col4 col5 col6
#> 1   63   NA 0.3327998  blue   NA   NA
#> 2    4   NA 0.5683527  blue   NA   NA
#> 3   34   NA 0.2522057 green   NA   NA
#> 4   35   NA 0.4640136 green   NA   NA
#> 5   89   NA 0.9176605  blue   NA   NA
#> 6   86   NA 0.9728442  blue   NA   NA